TXfilter and OHLC done

+++ OHLC.py
@@ -1,5 +1,14 @@
-# -*- coding: UTF-8 -*-
+##=============================
+#name:OHLC.py
+#argument:
+# input file:csv file
+#author:srhuang
+#email:lukyandy3162@gmail.com
+#=============================
+
+#===============
 #import section
+#===============
 import numpy as np
 import datetime
 import os
@@ -7,7 +16,10 @@ import sys
 import matplotlib.dates as mdates
 from os.path import exists
 
-#variables
+#================
+#variable section
+#================
+outputFolder="OHLC/"
 Cycle = 300
 OHLC=[]
 lastAmount=0
@@ -15,18 +27,21 @@ lowPrice=0
 highPrice=0
 date=0
 
-#argument
-file_name=sys.argv[1]
-outputFolder=sys.argv[2]
-output_file_name="OHLC"+"_"+file_name.split("_")[1].split(".")[0]+".csv"
+#=================
+#argument section
+#=================
+file_path=sys.argv[1]
+file_name=sys.argv[1].split('/')[1]
+output_file_name="OHLC"+"_"+file_name.split("_")[1]
 
-#functions
+#=================
+#function section
+#=================
 def TimetoNumber(time):
   time=time.zfill(6)
   sec=int(time[:2])*3600+int(time[2:4])*60+int(time[4:6])
   return sec
 
- #數值轉時間
 def NumbertoTime(sec):
  TS=str(sec%60).zfill(2)
  sec=sec/60
@@ -35,17 +50,33 @@ def NumbertoTime(sec):
  TH=str(sec%60).zfill(2)
  return TH+TM+TS
 
-#process begin
-print os.path.basename(__file__)
-print "Current Folder:"+os.getcwd()
-output_file = open(outputFolder+output_file_name,'w')
+#===============
+#progress start
+#===============
+print ">>>>> "+os.path.basename(__file__)+" "+sys.argv[1]
+#print "Current Folder:"+os.getcwd()
+
+
+#check the input file
+if exists(file_path):
+  input_file = open ( file_path,"r" )
+else:
+  print file_path+" is NOT exist."
+  sys.exit(0)
+
+#check and open the output file
+if exists(outputFolder+output_file_name):
+  print output_file_name + " is already exist."
+  sys.exit(0)
+else:
+  output_file = open(outputFolder+output_file_name,'w')
+
 #calculate start time and end time
 STime = TimetoNumber('084500')
 ETime = TimetoNumber('134500')
 
-#取得成交資訊
-fileHandle = open ( file_name,"r" )
-lineList = fileHandle.readlines()
+#calculate the OHLC and write into file
+lineList = input_file.readlines()
 for line in lineList:
 	#print line
   Input=line.replace(" ", "").split(",")
@@ -79,3 +110,5 @@ print "Lowest Price :", lowPrice, "Highest Price :", highPrice
 for data in OHLC:
   line=str(mdates.num2date(data[0]).time())+" "+str(data[1])+" "+str(data[2])+" "+str(data[3])+" "+str(data[4])+" "+str(data[5])+"\n"
   output_file.write(line)
+  
+print "<<<<< "+os.path.basename(__file__)+" done."
